requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
python-multipart==0.0.22
starlette==1.0.0
langchain==1.2.13
langchain-core==1.2.22
langchain-openai==1.1.12
langchain-chroma==1.1.0
langchain-text-splitters==1.1.1
langgraph==1.1.3
openai==2.30.0
tiktoken==0.12.0
chromadb==1.5.5
docling==2.82.0
docling-core==2.70.2
python-dotenv==1.2.2
pydantic==2.12.5
pydantic-settings==2.13.1
httpx==0.28.1
tenacity==9.1.4
PyYAML==6.0.3
python-dateutil==2.9.0.post0
tqdm==4.67.3